refactor(test package): 테스트 패키지 구성 변경
