--- src/test/java/kitchenpos/bo/InMemoryProductDao.java
@@ -1,26 +1,0 @@
-package kitchenpos.bo;
-
-import kitchenpos.product.domain.ProductDao;
-import kitchenpos.product.domain.Product;
-
-import java.util.*;
-
-public class InMemoryProductDao implements ProductDao {
-    private final Map<Long, Product> entities = new HashMap<>();
-
-    @Override
-    public Product save(final Product entity) {
-        entities.put(entity.getId(), entity);
-        return entity;
-    }
-
-    @Override
-    public Optional<Product> findById(final Long id) {
-        return Optional.ofNullable(entities.get(id));
-    }
-
-    @Override
-    public List<Product> findAll() {
-        return new ArrayList<>(entities.values());
-    }
-}
