+++ src/test/java/kitchenpos/menu/menuGroup/application/MenuGroupBoTest.java
@@ -1,0 +1,59 @@
+package kitchenpos.menu.menuGroup.application;
+
+import kitchenpos.menu.menuGroup.infra.InMemoryMenuGroupDao;
+import kitchenpos.menu.menuGroup.domain.MenuGroupDao;
+import kitchenpos.menu.menuGroup.domain.MenuGroup;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import java.util.Arrays;
+import java.util.List;
+
+import static kitchenpos.Fixtures.twoChickens;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class MenuGroupBoTest {
+    private final MenuGroupDao menuGroupDao = new InMemoryMenuGroupDao();
+
+    private MenuGroupBo menuGroupBo;
+
+    @BeforeEach
+    void setUp() {
+        menuGroupBo = new MenuGroupBo(menuGroupDao);
+    }
+
+    @DisplayName("메뉴 그룹을 등록할 수 있다.")
+    @Test
+    void create() {
+        // given
+        final MenuGroup expected = twoChickens();
+
+        // when
+        final MenuGroup actual = menuGroupBo.create(expected);
+
+        // then
+        assertMenuGroup(expected, actual);
+    }
+
+    @DisplayName("메뉴 그룹의 목록을 조회할 수 있다.")
+    @Test
+    void list() {
+        // given
+        final MenuGroup twoChickens = menuGroupDao.save(twoChickens());
+
+        // when
+        final List<MenuGroup> actual = menuGroupBo.list();
+
+        // then
+        assertThat(actual).containsExactlyInAnyOrderElementsOf(Arrays.asList(twoChickens));
+    }
+
+    private void assertMenuGroup(final MenuGroup expected, final MenuGroup actual) {
+        assertThat(actual).isNotNull();
+        assertAll(
+                () -> assertThat(actual.getName()).isEqualTo(expected.getName())
+        );
+    }
+}
