+++ src/test/java/kitchenpos/product/application/ProductBoTest.java
@@ -1,0 +1,81 @@
+package kitchenpos.product.application;
+
+import kitchenpos.product.infra.InMemoryProductDao;
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.product.domain.Product;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.NullSource;
+import org.junit.jupiter.params.provider.ValueSource;
+
+import java.math.BigDecimal;
+import java.util.Arrays;
+import java.util.List;
+
+import static kitchenpos.Fixtures.friedChicken;
+import static kitchenpos.Fixtures.seasonedChicken;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.junit.jupiter.api.Assertions.assertAll;
+
+class ProductBoTest {
+    private final ProductDao productDao = new InMemoryProductDao();
+
+    private ProductBo productBo;
+
+    @BeforeEach
+    void setUp() {
+        productBo = new ProductBo(productDao);
+    }
+
+    @DisplayName("상품을 등록할 수 있다.")
+    @Test
+    void create() {
+        // given
+        final Product expected = friedChicken();
+
+        // when
+        final Product actual = productBo.create(expected);
+
+        // then
+        assertProduct(expected, actual);
+    }
+
+    @DisplayName("상품의 가격이 올바르지 않으면 등록할 수 없다.")
+    @ParameterizedTest
+    @NullSource
+    @ValueSource(strings = "-1000")
+    void create(final BigDecimal price) {
+        // given
+        final Product expected = friedChicken();
+        expected.setPrice(price);
+
+        // when
+        // then
+        assertThatExceptionOfType(IllegalArgumentException.class).isThrownBy(() -> productBo.create(expected));
+    }
+
+    @DisplayName("상품의 목록을 조회할 수 있다.")
+    @Test
+    void list() {
+        // given
+        final Product friedChicken = productDao.save(friedChicken());
+        final Product seasonedChicken = productDao.save(seasonedChicken());
+
+        // when
+        final List<Product> actual = productBo.list();
+
+        // then
+        assertThat(actual).containsExactlyInAnyOrderElementsOf(Arrays.asList(friedChicken, seasonedChicken));
+    }
+
+    private void assertProduct(final Product expected, final Product actual) {
+        assertThat(actual).isNotNull();
+        assertAll(
+                () -> assertThat(actual.getName()).isEqualTo(expected.getName()),
+                () -> assertThat(actual.getPrice()).isEqualTo(expected.getPrice())
+        );
+    }
+}
