+++ src/test/java/kitchenpos/product/infra/InMemoryProductDao.java
@@ -1,0 +1,26 @@
+package kitchenpos.product.infra;
+
+import kitchenpos.product.domain.ProductDao;
+import kitchenpos.product.domain.Product;
+
+import java.util.*;
+
+public class InMemoryProductDao implements ProductDao {
+    private final Map<Long, Product> entities = new HashMap<>();
+
+    @Override
+    public Product save(final Product entity) {
+        entities.put(entity.getId(), entity);
+        return entity;
+    }
+
+    @Override
+    public Optional<Product> findById(final Long id) {
+        return Optional.ofNullable(entities.get(id));
+    }
+
+    @Override
+    public List<Product> findAll() {
+        return new ArrayList<>(entities.values());
+    }
+}
